Added semantic actions to grammar.
Creates a root expression (extern variable in absyn.h) and
is used in parse.c.
Can use prabsyn.[ch] to print out the abstract syntax tree.

Needed to create a nonterminal for an identifier, so that we only get
the identifier in the string passed to S_Symbol. Also, required to modify
the symbol constructor function to allocate a unique copy of the
identifier string, otherwise the parser will eat the entire input into the
identifer string -- including terminals. The nonterminal has the lowest
precedence when used with lvalue; to get rid of a shift/reduce conflict
with the array exp lvalue production.

Added a new rule for a records initialisation field because it uses a
different list structure; efieldlist instead of explist which arglist
uses.

Modified makefile all rule to create AST generating parser (parsertest has
its own rule).