Semantic Analysis mostly completed and tested.
All tests check correctly, some tidying up of the semantic module may be
necessary later but since I will be augmenting this later for activation
records and translation to the IR.

One concern is how to prevent the for loop iterator variable from being
modified inside the body of the for loop. This may be addressable in later
stages.