Added frame and translate implementation modules.
The implementation is for the AMD64 architecture. Documentation will be
added (in the form of an ABI document) if that is permitted by the source
of the reference material.

The environment module has been altered; the data structures for variable
entrys in the value environment have been augmented with the translate
module's access and level structures. Although, the implementation does
not support these additions (should compile since the interface to these
structures has not been altered).

Implementation does not calculate escaping variables (will defer this to
later date).