Added Tree module that is the IR for the compiler.
Updated Translate module to make use of the Tree module (as well as the
Frame implementation).

The Frame implementation now handles the case of non-escaping parameters;
up to the limit provided by the target machine. After this limit has been
reached, the implmentation then proceeds to allocate parameters on the
stack frame. The implementation does not preserve space on the stack frame
for the parameters allocated to registers (in case the function needs to use the
registers for another function call and so needs to save the parameters on
the frame before the call, for example). This may be looked at a later
date.

Semantic module altered to make new level upon function declaration but
function entry and variable expressions have not been augmented yet.
Need to calculate static links when translating a variable access so we get the
correct location of the variable.

Added constructors/destructors for the three kinds of expressions
represented in the Tree intermediate representation: expressions that
yield a value; expressions that do not return a value and conditional
expressions.

I have also added a module that prints a tree (provided by Appel) and an
interface (although no implementation as yet) for an Escape module which
calculates the whether a variable escapes and records this information in
the abstract syntax.