Added translation expression lists to translate.c
Translation expression lists are used to collect the arguments for a
function call. The list is an abstract data type and has an interface as
specified in the translate module's header file. In this interface, is a
conversion function that converts the translation list to a tree IR
expression list.

Tree expression lists are used in the function call mechanism for passing
arguments and the static link.